Cleanup crypter interface
