+++ encryption-service/impl/crypt/aes_test.go
@@ -27,7 +27,7 @@ var aad, _ = hex.DecodeString("feedfacedeadbeeffeedfacedeadbeefabaddad2")
 var plaintext, _ = hex.DecodeString("d9313225f88406e5a55909c5aff5269a86a7a9531534f7da2e4c303d8a318a721c3c0c95956809532fcf0e2449a6b525b16aedf5aa0de657ba637b39")
 var ciphertextHex = "522dc1f099567d07f47f37a32a84427d643a8cdcbfe5c0c97598a2bd2555d1aa8cb08e48590dbb3da7b08b1056828838c5f61e6393ba7a0abcc9f66276fc6ece0f4e1768cddf8853bb2d551b" + hex.EncodeToString(nonce)
 
-func TestEncryptPilot(t *testing.T) {
+func TestEncryptDecrypt(t *testing.T) {
 	crypter := &AESCrypter{}
 
 	tmpReader := rand.Reader
@@ -48,7 +48,7 @@ func TestEncryptPilot(t *testing.T) {
 
 	gotPlaintext, err := crypter.Decrypt(ciphertext, aad, oek)
 	if err != nil {
-		t.Fatalf("DecryptPilot: %v", err)
+		t.Fatalf("Decrypt: %v", err)
 	}
 	if !bytes.Equal(plaintext, gotPlaintext) {
 		t.Fatalf("plaintext doesn't match:\n%x\n%x\n", plaintext, gotPlaintext)
